Added chips "dot_tuning_manager" and "sketch_refresher" to the gui motherboard. This is the first step for transforming the sketch refreshing in a side effect of the gui motherboard stabilization.
All dot tuning toolbar widgets, except cable revert, have a callback that set a
related wire (introduced in the class Mariokit.Dotoptions.network). This action
provokes a system stabilization, then a sketch refreshing.

There aren't now direct calls to st#refresh_sketch() in the file 
gui_toolbar_DOT_TUNING.ml, except in the "cable revert" callback. 
Actually, I'm now not able to easily change this behaviour because the 
information "cable reverted" is stored into the cable object, which is a dinamic
part of the state. I need before an enhanced support of dinamic systems in chip.ml
(specially for the problem of garbage).