Improved the portability of marionnet's projects. A configuration variable "MTIME" specifies now the modification-time of the related filesystem. Because of a user-mode-linux's protection mechanism, this setting is very important for sharing marionnet's projects between different installations. Actually, if we open on a computer a project created elsewhere, the involved filesystems must be the same, of course, but they must also have exactly the same MTIME. For this reason, it's a good idea to run the following post-installation command:
sudo touch -d $(date -d "@$MTIME") $FILESYSTEM

where $FILESYSTEM is the location of the filesystem image the disk. On the other hand, when we are providing a new filesystem, we can set this variable (once our filesystem is completely finished and stable) simply running the command:

stat -c "%Y" $FILESYSTEM 

Marionnet gives now a warning (at loading-time) when the installation contains a filesystem (backend) with an "mtime" different from the value announced by this variable. This warning suggests also a solution. Here an example:
---
(TITLE:) 
Modification time (MTIME) inconsistency

(MESSAGE:)
The filesystem `debian-lenny-sid-2008' has the mtime 1291051963, but the expected value was 1291051964.
Please run the command:

sudo touch -d $(date -d @1291051964) /usr/local/share/marionnet/filesystems/machine-debian-lenny-sid-2008

in order to fix this inconsistency. Otherwise, machines with this filesystem defined in a project created elsewhere can not be restarted.
---